GamePacket 타입 분할
