+++ Assets/Scripts/Net/GamePackets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8a7c9c9426205b047b1724c44430ebce
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
