+++ Assets/Scripts/Net/GamePackets/GameBuffGeneratePacket.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8a30c1a05edcce842a32081028cce92e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
