+++ Assets/Scripts/Net/GamePackets/GameBuffReleasePacket.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1a9c92586b3807d439aabc664c198f1a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
