+++ Assets/Scripts/Net/GamePackets/GameFragmentAreaPacket.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: aed908ce76f5a2847ac054062a6c6bf1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
