+++ Assets/Scripts/Net/GamePackets/GameParticlePacket.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8027e43ee6ed1494b81e4ade484a817c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
