+++ Assets/Scripts/Net/GamePackets/PacketSerializer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8c4b109b4e84b294da5c1ed8afdbee39
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
